Created standalone 'utils.py' file
