+++ algopytest/__init__.py
@@ -31,6 +31,7 @@ from .client_ops import (
     compile_program,
     suggested_params,
 )
+from .entities import AlgoUser
 from .smart_program_ops import deploy_smart_contract
 from .transaction_ops import (
     call_app,
@@ -50,6 +51,7 @@ from .transaction_ops import (
 
 # The functions to expose for sphinx documentation
 __all__ = [
+    "AlgoUser",
     "deploy_smart_contract",
     "account_balance",
     "application_global_state",
