+++ algopytest/client_ops.py
@@ -16,7 +16,6 @@ from typing import Any, Callable, Optional
 
 import pyteal
 from algosdk import mnemonic
-from algosdk.encoding import encode_address
 from algosdk.error import IndexerHTTPError
 from algosdk.future import transaction
 from algosdk.future.transaction import LogicSig, PaymentTxn, wait_for_confirmation
@@ -25,6 +24,7 @@ from pyteal import Mode, compileTeal
 
 from .config_params import ConfigParams
 from .entities import AlgoUser
+from .utils import _convert_algo_dict
 
 
 ## CLIENTS
@@ -233,7 +233,6 @@ def account_balance(account: AlgoUser) -> int:
     return account_data["amount"]
 
 
-## UTILITY
 def _compile_source(source: str) -> bytes:
     """Compile and return teal binary code."""
     compile_response = _algod_client().compile(source)
@@ -259,44 +258,3 @@ def compile_program(program: pyteal.Expr, mode: Mode, version: int = 5) -> bytes
     """
     source = compileTeal(program, mode=mode, version=version)
     return _compile_source(source)
-
-
-def logic_signature(teal_source: Any) -> Any:
-    """Create and return logic signature for provided `teal_source`."""
-    # TODO: Typing and general behavior
-    compiled_binary = _compile_source(teal_source)
-    return LogicSig(compiled_binary)
-
-
-def _base64_to_str(b64: str) -> str:
-    """Converts a b64 encoded string to a normal UTF-8 string."""
-    # Decode the base64 to bytes and then decode them as a UTF-8 string
-    byte_decoding = base64.b64decode(b64)
-    return byte_decoding.decode("utf-8")
-
-
-def _convert_algo_dict(
-    algo_dict: list[dict[str, Any]], address_fields: Optional[list[str]]
-) -> dict[str, str]:
-    """Converts an Algorand dictionary to a Python one."""
-    # Materialize the `address_fields` to a list type
-    address_fields = address_fields or []
-
-    ret = {}
-    for entry in algo_dict:
-        key = _base64_to_str(entry["key"])
-
-        value_type = entry["value"]["type"]
-
-        if value_type == 1 and key not in address_fields:  # Bytes non-address
-            value = _base64_to_str(entry["value"]["bytes"])
-        elif value_type == 1 and key in address_fields:  # Bytes address
-            value = encode_address(base64.b64decode(entry["value"]["bytes"]))
-        elif value_type == 2:  # Integer
-            value = entry["value"]["uint"]
-        else:
-            raise ValueError(f"Unknown value type for key: {key}")
-
-        ret[key] = value
-
-    return ret
